+++ src/renderer/components/LoginScreen.tsx
@@ -1,21 +1,11 @@
-import React, { useState, useEffect, Fragment, useContext } from 'react'
-import { sendToBackend, ipcBackend } from '../ipc'
+import React, { useState, useEffect, useContext } from 'react'
+import { ipcBackend } from '../ipc'
 import { Credentials } from '../../shared/shared-types'
 import LoginForm, {
   defaultCredentials,
   ConfigureProgressDialog,
 } from './LoginForm'
-import {
-  Button,
-  Classes,
-  Elevation,
-  Intent,
-  Card,
-  Alignment,
-  Navbar,
-  NavbarGroup,
-  NavbarHeading,
-} from '@blueprintjs/core'
+import { Classes, Elevation, Intent, Card, Icon } from '@blueprintjs/core'
 import { DeltaProgressBar } from './Login-Styles'
 import { getLogger } from '../../shared/logger'
 import { ScreenContext, useTranslationFunction } from '../contexts'
@@ -35,11 +25,7 @@ import { DeltaBackend } from '../delta-remote'
 import { Screens } from '../ScreenController'
 import { IpcRendererEvent } from 'electron'
 import { Avatar } from './Avatar'
-import {
-  PseudoListItemAddContact,
-  PseudoListItem,
-} from './helpers/PseudoListItem'
-import { ContactListItem } from './contact/ContactListItem'
+import { PseudoContact } from './contact/Contact'
 
 const log = getLogger('renderer/components/LoginScreen')
 
@@ -162,7 +148,7 @@ export default function LoginScreen({
   loadAccount: (account: DeltaChatAccount) => {}
 }) {
   const tx = useTranslationFunction()
-  const { openDialog, changeScreen } = useContext(ScreenContext)
+  const { openDialog } = useContext(ScreenContext)
 
   const [credentials, setCredentials] = useState<Credentials>(
     defaultCredentials()
@@ -184,21 +170,6 @@ export default function LoginScreen({
       loadAccount(account)
     }
     openDialog(ConfigureProgressDialog, { credentials, onSuccess })
-  }
-
-  const removeAccount = (login: DeltaChatAccount) => {
-    const message = tx('forget_login_confirmation_desktop')
-    openDialog('ConfirmationDialog', {
-      message,
-      confirmLabel: tx('delete_account'),
-      isConfirmDanger: true,
-      cb: async (yes: boolean) => {
-        if (yes) {
-          await DeltaBackend.call('login.forgetAccount', login)
-          refreshAccounts()
-        }
-      },
-    })
   }
 
   if (logins === null) return null
@@ -271,26 +242,9 @@ export default function LoginScreen({
                     />
                     <DeltaDialogBody>
                       <DeltaDialogContent noPadding={true}>
-                        <div className='accounts'>
-                          <ul>
-                            <PseudoListItem
-                              id='action-go-to-login'
-                              cutoff='+'
-                              text={tx('add_account')}
-                              onClick={() => setView('login')}
-                            />
-                            {logins.map(
-                              (login: DeltaChatAccount, index: Number) => (
-                                <AccountItem
-                                  key={`login-${index}`}
-                                  login={login}
-                                  loadAccount={loadAccount}
-                                  removeAccount={removeAccount}
-                                />
-                              )
-                            )}
-                          </ul>
-                        </div>
+                        <AccountSelection
+                          {...{ refreshAccounts, setView, logins, loadAccount }}
+                        />
                       </DeltaDialogContent>
                     </DeltaDialogBody>
                   </>
@@ -312,37 +266,142 @@ export default function LoginScreen({
   )
 }
 
-export function AccountItem({
+function AccountSelection({
+  refreshAccounts,
+  setView,
+  logins,
+  loadAccount,
+}: {
+  refreshAccounts: () => Promise<void>
+  setView: React.Dispatch<React.SetStateAction<string>>
+  logins: any
+  loadAccount: (account: DeltaChatAccount) => void
+}) {
+  const tx = useTranslationFunction()
+  const { openDialog } = useContext(ScreenContext)
+
+  const removeAccount = (login: DeltaChatAccount) => {
+    const header = tx('delete_account_confirmation_header_desktop', login.addr)
+    const message = tx('delete_account_confirmation_desktop', login.addr)
+    openDialog('ConfirmationDialog', {
+      header,
+      message,
+      confirmLabel: tx('delete_account'),
+      isConfirmDanger: true,
+      cb: async (yes: boolean) => {
+        if (yes) {
+          await DeltaBackend.call('login.forgetAccount', login)
+          refreshAccounts()
+        }
+      },
+    })
+  }
+
+  useEffect(() => {
+    const onKeyDown = (ev: KeyboardEvent) => {
+      const parent = document.querySelector<HTMLDivElement>('#accounts')
+      const current = parent?.querySelector(':focus')
+
+      if (ev.key == 'ArrowDown') {
+        if (current && current.nextElementSibling) {
+          ;(current.nextElementSibling as HTMLDivElement)?.focus()
+        } else {
+          ;(parent?.firstElementChild as HTMLDivElement).focus()
+        }
+      } else if (ev.key == 'ArrowUp') {
+        if (current && current.previousElementSibling) {
+          ;(current.previousElementSibling as HTMLDivElement)?.focus()
+        } else {
+          ;(parent?.lastElementChild as HTMLDivElement).focus()
+        }
+      } else if (ev.key == 'Enter') {
+        if (current) {
+          ;(current as HTMLDivElement)?.click()
+        }
+      }
+    }
+
+    document.addEventListener('keydown', onKeyDown)
+    return () => document.removeEventListener('keydown', onKeyDown)
+  })
+
+  return (
+    <div className='accounts' id='accounts' role='menu'>
+      <div
+        role='menu-item'
+        id='action-go-to-login'
+        className='contact-list-item'
+        onClick={() => setView('login')}
+        tabIndex={0}
+      >
+        <PseudoContact cutoff='+' text={tx('add_account')}></PseudoContact>
+      </div>
+      {logins.map((login: DeltaChatAccount, index: Number) => (
+        <AccountItem
+          key={`login-${index}`}
+          login={login}
+          loadAccount={loadAccount}
+          removeAccount={removeAccount}
+        />
+      ))}
+    </div>
+  )
+}
+
+function AccountItem({
   login,
   loadAccount,
   removeAccount,
 }: {
   login: DeltaChatAccount
-  loadAccount: todo
-  removeAccount: todo
+  loadAccount: (account: DeltaChatAccount) => void
+  removeAccount: (account: DeltaChatAccount) => void
 }) {
-  const contact = {
-    address: login.addr,
-    color: login.color,
-    displayName: login.displayname || login.addr,
-    firstName: '',
-    id: 0,
-    name: login.displayname,
-    profileImage: login.profileImage,
-    nameAndAddr: login.displayname,
-    isBlocked: false,
-    isVerified: false,
+  const removeAction = (ev: React.MouseEvent<HTMLDivElement, MouseEvent>) => {
+    ev?.stopPropagation()
+    removeAccount(login)
   }
 
+  const title = window.static_translate('account_info_hover_tooltip_desktop', [
+    login.addr,
+    filesizeConverter(login.size),
+    'accounts' + login.path.split('accounts')[1],
+  ])
+
   return (
-    <ContactListItem
-      key={login.addr}
-      contact={contact}
+    <div
+      role='menu-item'
+      className='contact-list-item'
       onClick={() => loadAccount(login)}
-      showCheckbox={false}
-      checked={false}
-      showRemove={true}
-      onRemoveClick={() => removeAccount(login)}
-    />
+      tabIndex={0}
+    >
+      <div className='contact'>
+        <Avatar
+          {...{
+            avatarPath: login.profileImage,
+            color: login.color,
+            displayName: login.displayname,
+          }}
+        />
+        <div className='contact-name'>
+          <div className='display-name'>{login.displayname || login.addr}</div>
+          <div
+            className='email'
+            style={{ display: 'inline-block' }}
+            title={title}
+          >
+            {login.addr}
+          </div>
+        </div>
+      </div>
+      <div
+        role='button'
+        aria-label={window.static_translate('delete_account')}
+        className='remove-icon'
+        onClick={removeAction}
+      >
+        <Icon icon='cross' />
+      </div>
+    </div>
   )
 }